add sam update

+++ optimizers/darts/utils.py
@@ -186,11 +186,12 @@ def drop_path(x, drop_prob):
 
 def create_exp_dir(path, scripts_to_save=None):
     if not os.path.exists(path):
-        os.makedirs(path)
+        os.mkdir(path)
     print('Experiment dir : {}'.format(path))
 
-    if scripts_to_save is not None:
-        os.mkdir(os.path.join(path, 'scripts'))
+    script_path = os.path.join(path, 'scripts')
+    if not os.path.exists(script_path) and scripts_to_save is not None:
+        os.mkdir(script_path)
         for script in scripts_to_save:
             dst_file = os.path.join(path, 'scripts', os.path.basename(script))
             shutil.copyfile(script, dst_file)
